+++ app.py
@@ -83,6 +83,7 @@ NAV_PAGES = ["Home", "Dataset", "Volume", "Pipeline", "Detection", "Lineage", "E
 
 st.markdown(APP_CSS, unsafe_allow_html=True)
 
+
 init_session()